Built an API interactor and a trades to OHLC convertor
The API interactor was meant to pull data from the Kraken and VALR exchanges but I managed to get the data via email rather.

The trades to OHLC convertor reads the every trade the happens in a dataset from a staring time and creates a dataset of Open, High, Low, Close, Volume and number of Trades from it.